Thêm bcryptpass, chia role manager-resident cho user

+++ src/main/java/com/example/apartmentmanagement/repository/ResidentRepository.java
@@ -1,13 +1,13 @@
 package com.example.apartmentmanagement.repository;
 
 import com.example.apartmentmanagement.entity.Resident;
+import com.example.apartmentmanagement.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
-public interface ResidentRepository extends JpaRepository<Resident, Integer> {
-    List<Resident> findByApartmentId(int apartmentId);
+public interface ResidentRepository extends JpaRepository<Resident, Long> {
+    List<Resident> findByApartmentId(Long apartmentId);
 }
